+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
+- Resolved Wiktionary language names for languages with at least 100
+  pronunciation entries. (#52)
+
 ### Changed
 ### Deprecated
 ### Removed